% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\title{boot}
\usage{
boot()
}
\description{
Cleans global environment and sets working directory
}
\details{
This function prepares any script or RMarkdown file to be ran from scripts directory
with a clean global environment. This avoids issues with previously existing variables in the
global environment and RStudio's knitting not using the project's working directory.
}
\author{
Ricardo Vladimiro
}

