\name{ognd.tools-package}
\alias{ognd.tools-package}
\alias{ognd.tools}
\docType{package}
\title{
\packageTitle{ognd.tools}
}
\description{
\packageDescription{ognd.tools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ognd.tools}
\packageIndices{ognd.tools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ognd.tools}

Maintainer: \packageMaintainer{ognd.tools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
