% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_strings.R
\name{random_strings}
\alias{random_strings}
\title{random_strings}
\usage{
random_strings(n = NULL, length = 8, include_numbers = T)
}
\arguments{
\item{n}{The number of random strings}

\item{length}{The length of each individual string}

\item{include_numbers}{If true, numbers are used to create the string}
}
\value{
A vector of n random strings of length length
}
\description{
Generates random strings
}
\details{
The random string generation will generate a vector of n random strings, created by
sampling from upper and lower letters. If include_numbers is true, then numbers will also be
used to build the random strings. Random strings are created taking into consideration a
random seed if it is set by set.seed() function.
}
\author{
Ricardo Vladimiro
}
\seealso{
set.seed()
}

